프론트 요청 리팩토링 및 게시글 세부사항 반환 (#31)
* feat : 리뷰 ,자료실 좋아요 테이블 생성

* feat : myPost 게시판 스프링시큐리티 적용

* feat : 리뷰 , 자료실 좋아요 누르는 기능API 생성

* feat : 자료실 인기순 페이징 API 호출

* refactor : post 요청 저장시 id반환 변경

* feat : 게시글 세부사항 API

+++ src/main/java/kusitms/gallae/domain/Review.java
@@ -41,7 +41,7 @@ public class Review {
 
     private String hashtag;
 
-    private int likes;
+    private Long likes;
 
     @Column
     @CreationTimestamp
